+++ NEWS.md
@@ -1,9 +1,14 @@
+---
+output: html_document
+---
 trimr 1.0.0.9000
 ===========
-
+* Updated modifiedRecursive function to avoid bug if only two response times
+   are passed to the function. (Thanks to Ayala Allon for noticing this bug.)
 
 trimr 1.0.0
 ===========
+* CRAN release
 * Vignettes added